semacomm-master/requirements.txt
deepspeed==0.17.5
diffusers==0.35.1
einops==0.8.1
fastapi==0.116.1
fsspec==2024.3.1
huggingface_hub==0.34.4
lpips==0.1.4
matplotlib==3.5.1
mindspore==2.7.0
mup==1.0.0
numpy==1.21.5
omegaconf==2.3.0
pandas==2.3.2
Pillow==11.3.0
psutil==5.9.6
pydantic==2.11.7
pyldpc==0.7.9
safetensors==0.6.2
scipy==1.8.0
seaborn==0.13.2
torchvision==0.17.2
uvicorn==0.35.0
wandb==0.21.3
xformers==0.0.32.post2
